% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{DistancesOnTorus}
\alias{DistancesOnTorus}
\title{Distances between matched points on a torus}
\usage{
DistancesOnTorus(dimension, linear.system.size, coords.1, coords.2)
}
\arguments{
\item{dimension}{Dimension of the simulation box}

\item{linear.system.size}{Linear system size L of the simulation box}

\item{coords.1}{Matrix with coordinates of points in 1st pattern (each row stores the coordinates of one point)}

\item{coords.2}{Matrix with coordinates of points in 2nd pattern (each row stores the coordinates of one point)}
}
\value{
Vector of distances, each row corresponds to the distances of the
 corresponding rows in coords.1 and coords.2
}
\description{
Output of distances on torus between points in two patterns (auxiliary function)
}
