% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{Lattice}
\alias{Lattice}
\title{Coordinates of a lattice}
\usage{
Lattice(dimension,lin.N,lin.len)
}
\arguments{
\item{dimension}{Dimension of the simulation box}

\item{lin.N}{Vector of linear numbers of lattice points}

\item{lin.len}{Lattice constant "a" (nearest neighbor distance)}
}
\value{
Matrix R^{prod(lin.N) x dimension}, each row contains coordinates of one cell center
}
\description{
Creates a patch of the lattice a*Z^d, where d is the dimension (auxiliary function)
}
