% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{MatchingInfileToInfile}
\alias{MatchingInfileToInfile}
\title{Stable matching of point patterns from two infiles}
\usage{
MatchingInfileToInfile(linear.system.size, infile.1, infile.2, fileout=FALSE, filename="example-matchedpp.dat", verbose=TRUE)
}
\arguments{
\item{linear.system.size}{Linear system size L of the simulation box [0,L)^d}

\item{infile.1}{Name of a file that contains a matrix, where each row stores the coordinates of one point}

\item{infile.2}{Name of another file that contains a matrix, where each row stores the coordinates of one point}

\item{fileout}{Output of matched points to file if set to TRUE (default=FALSE)}

\item{filename}{Name of output file (default="example-matchedpp.dat")}

\item{verbose}{Output of progress if set to TRUE (default=TRUE)}
}
\value{
Matrix that stores all matched points
        of the point pattern from infile.1 (in columns 1 to dimension)
        and of the point pattern from infile.2 (in columns dimension+1 to end).
        Each row contains one pair of matched points.
}
\description{
Stable matching of point patterns from two infiles
}
\examples{
tmp.bino <- "example-bpp.dat"
tmp.pois <- "example-ppp.dat"
linear.sys.size <- 4.0
n.pts <- 25
coords.bino <- matrix(runif(2*n.pts,0,linear.sys.size), nrow = n.pts)
n.pts <- rpois(1,n.pts)
coords.pois <- matrix(runif(2*n.pts,0,linear.sys.size), nrow = n.pts)
write.table(coords.bino,file=tmp.bino,row.names=FALSE, col.names=FALSE)
write.table(coords.pois,file=tmp.pois,row.names=FALSE, col.names=FALSE)
out <- MatchingInfileToInfile(linear.sys.size, tmp.pois, tmp.bino)
}
