% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{MatchingInfileToLattice}
\alias{MatchingInfileToLattice}
\title{Stable matching of points from an infile to a randomized lattice}
\usage{
MatchingInfileToLattice(dimension, linear.system.size, infile, fileout=FALSE, filename="example-matchedpp.dat", verbose=TRUE)
}
\arguments{
\item{dimension}{Dimension d of the simulation box}

\item{linear.system.size}{Linear system size L of the simulation box [0,L)^d}

\item{infile}{Name of file that contains a matrix, where each row stores the coordinates of one point}

\item{fileout}{Output of matched points to file if set to TRUE (default=FALSE)}

\item{filename}{Name of output file (default="example-matchedpp.dat")}

\item{verbose}{Output of progress if set to TRUE (default=TRUE)}
}
\value{
Matrix that stores all matched points
        of the point pattern from infile (in columns 1 to dimension)
        and of the lattice (in columns dimension+1 to end).
        Each row contains one pair of matched points.
}
\description{
Stable matching of points from an infile to a randomized lattice
}
\examples{
tmp.infile <- "example-ppp.dat"
linear.sys.size <- 4.0
n.pts <- rpois(1,1000.0*linear.sys.size^2)
coords.pois <- matrix(runif(2*n.pts,0,linear.sys.size), nrow = n.pts)
write.table(coords.pois,file=tmp.infile,row.names=FALSE, col.names=FALSE)
out <- MatchingInfileToLattice(dimension=2, linear.sys.size, tmp.infile)
}
