% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{MatchingPoissonToLattice}
\alias{MatchingPoissonToLattice}
\title{Stable matching of point patterns}
\usage{
MatchingPoissonToLattice(dimension, linear.system.size, intensity, fileout=FALSE, filename="example-matchedpp.dat", unthinned.fileout=FALSE, unthinned.filename="example-ppp.dat", verbose=TRUE)
}
\arguments{
\item{dimension}{Dimension d of the simulation box}

\item{linear.system.size}{Linear system size L of the simulation box [0,L)^d}

\item{intensity}{Intensity of the Poisson point process (mean number of points in[0,1)^d)}

\item{fileout}{Output of matched points to file if set to TRUE (default=FALSE)}

\item{filename}{Name of output file (default="example-matchedpp.dat")}

\item{unthinned.fileout}{Output of unthinned Poisson point process to file if set to TRUE (default=FALSE)}

\item{unthinned.filename}{Name of output file for the unthinned Poisson point process (default="example-ppp.dat")}

\item{verbose}{Output of progress if set to TRUE (default=TRUE)}
}
\value{
Matrix that stores all matched points
        of the Poisson point process (in columns 1 to dimension)
        and of the lattice (in columns dimension+1 to end).
        Each row contains one pair of matched points.
}
\description{
Stable matching of a randomized lattice to a Poisson point process
}
\details{
Mutual nearest neighbor matching of two point patterns on the torus.
imports RANN::n2
Stable matching of a randomized lattice to a Poisson point process
}
\examples{
out.1 <- MatchingPoissonToLattice(dimension=3,linear.system.size=4.0,intensity=25.0)
out.2 <- MatchingPoissonToLattice(2,10.0,2.5,fileout=TRUE,filename="example-2D-matched-possion-to-Z.dat")
}
