% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{MatchingPP}
\alias{MatchingPP}
\title{Stable matching of point patterns}
\usage{
MatchingPP(linear.system.size, coords.1, coords.2, fout=FALSE, fname="example-matchedpp.dat", verb=TRUE, nn2.treetype = "kd", nn2.eps=0)
}
\arguments{
\item{linear.system.size}{Linear system size L of the simulation box [0,L)^d}

\item{coords.1}{A point pattern stored in a matrix, where each row stores the coordinates of one point}

\item{coords.2}{Another point pattern stored in a matrix, where each row stores the coordinates of one point}

\item{fout}{Output of matched points to file if set to TRUE (default=FALSE)}

\item{fname}{Name of output file (default="example-matchedpp.dat")}

\item{verb}{Output of progress if set to TRUE (default=TRUE)}

\item{nn2.treetype}{Optional argument of the function nn2 from the package RANN, which is used for the nearest neighbor search: "Either the standard kd tree or a bd (box-decomposition, AMNSW98) tree which may perform better for larger point sets" (default="kd")}

\item{nn2.eps}{Optional argument of the function nn2 from the package RANN, which is used for the nearest neighbor search: "error bound: default of 0.0 implies exact nearest neighbour search" (default=0)}
}
\value{
Matrix that stores all matched points
        of the 1st point pattern (in columns 1 to dimension)
        and of the 2nd point pattern (in columns dimension+1 to end).
        Each row contains one pair of matched points.
}
\description{
Stable matching of two point patterns in a cubic simulation box [0,L)^d:
             It is the core function of the package MatchingPP.

             The nearest neighbor search uses the function nn2 from the package RANN.
}
\examples{
linear.sys.size <- 4.0
n.pts <- 25
coords.bino <- matrix(runif(2*n.pts,0,linear.sys.size), nrow = n.pts)
n.pts <- rpois(1,n.pts)
coords.pois <- matrix(runif(2*n.pts,0,linear.sys.size), nrow = n.pts)
out <- MatchingPP(linear.sys.size, coords.pois, coords.bino)
}
