% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingPP.R
\name{PeriodicCopies}
\alias{PeriodicCopies}
\title{Create periodic copies of points in a simulation box}
\usage{
PeriodicCopies(dimension,coordinates,L)
}
\arguments{
\item{dimension}{Dimension of the simulation box}

\item{coordinates}{Matrix that stores the coordinates of the points}

\item{L}{Linear size of cubic simulation box}
}
\value{
Matrix that stores all periodic copies
}
\description{
Create 3^d periodic copies of points in [0,L)^d (auxiliary function)
}
